lambda-sentinel-data/requirements.txt
awslambdaric
boto3==1.35.86
botocore==1.35.86
certifi==2024.12.14
charset-normalizer==3.4.0
idna==3.10
jmespath==1.0.1
joblib==1.4.2
pygdal==3.6.3.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
s3transfer==0.10.4
scikit-learn==1.6.0
scipy==1.14.1
six==1.17.0
threadpoolctl==3.5.0
urllib3==2.3.0
numpy<2.0.0